Created a common utils
