+++ packages/selenium-ide/src/common/__mocks__/utils.js
