+++ packages/selenium-ide/src/content/utils.js
@@ -16,7 +16,6 @@
 // under the License.
 
 import SeleniumError from "./SeleniumError";
-import parser from "ua-parser-js";
 
 /**
  * Parses a Selenium locator, returning its type and the unprefixed locator
@@ -104,23 +103,7 @@ export function upperFirstChar(str) {
   return str.charAt(0).toUpperCase() + str.substr(1);
 }
 
-export const isProduction = process.env.NODE_ENV === "production";
-
 export function calculateFrameIndex(indicatorIndex, targetFrameIndex) {
   if (indicatorIndex < 0) return targetFrameIndex;
   return (indicatorIndex < targetFrameIndex) ? targetFrameIndex - 1 : targetFrameIndex;
-}
-
-export const isStaging = process.env.NODE_ENV === "staging";
-
-export const isTest = process.env.NODE_ENV === "test";
-
-export const userAgent = parser(window.navigator.userAgent);
-
-export function isChrome() {
-  userAgent.browser.name === "Chrome";
-}
-
-export function isFirefox() {
-  userAgent.browser.name === "Firefox";
 }
